+++ python/test/pymidibridge/test_protocol_parallel.py
@@ -29,7 +29,7 @@ class TestProtocolParallel(unittest.TestCase):
 
         bridge_send = PyMidiBridge(
             midi = midi_send,
-            storage = storage_send,
+            storage_factory = get_mock_storage_factory(storage_send),
             event_handler = MockEventHandler()
         )
 
@@ -39,7 +39,7 @@ class TestProtocolParallel(unittest.TestCase):
 
         bridge_receive = PyMidiBridge(
             midi = midi_receive,
-            storage = storage_receive,
+            storage_factory = get_mock_storage_factory(storage_receive),
             event_handler = MockEventHandler()
         )