+++ python/test/pymidibridge/test_protocol_timeout.py
@@ -31,7 +31,7 @@ class TestProtocolTimeout(unittest.TestCase):
 
         bridge_send = PyMidiBridge(
             midi = midi_send,
-            storage = storage_send,
+            storage_factory = get_mock_storage_factory(storage_send),
             event_handler = MockEventHandler()
         )
 
@@ -41,7 +41,7 @@ class TestProtocolTimeout(unittest.TestCase):
 
         bridge_receive = PyMidiBridge(
             midi = midi_receive,
-            storage = storage_receive,
+            storage_factory = get_mock_storage_factory(storage_receive),
             event_handler = MockEventHandler()
         )
 
@@ -110,7 +110,7 @@ class TestProtocolTimeout(unittest.TestCase):
 
         bridge_send = PyMidiBridge(
             midi = midi_send,
-            storage = storage_send,
+            storage_factory = get_mock_storage_factory(storage_send),
             event_handler = MockEventHandler()
         )
 
@@ -120,7 +120,7 @@ class TestProtocolTimeout(unittest.TestCase):
 
         bridge_receive = PyMidiBridge(
             midi = midi_receive,
-            storage = storage_receive,
+            storage_factory = get_mock_storage_factory(storage_receive),
             event_handler = MockEventHandler()
         )
 
@@ -173,7 +173,7 @@ class TestProtocolTimeout(unittest.TestCase):
         other_midi = MockMidiSender()
         other_bridge = PyMidiBridge(
             midi = other_midi, 
-            storage = MockStorageProvider(),
+            storage_factory = get_mock_storage_factory(),
             event_handler = events
         )