+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# MidiBridge 0.5.3
+Performance optimizations for python:
+- Late importing of bridge and storage providers at time the first message arrives
+- Pre-Compiled mpy versions
+
 # MidiBridge 0.5.2
 Minor bug fixes:
 - Removed warning for syntax errors on save